Finished front-end refactor
- moved `lbalConfig.hpp` to `details/`
- made supporting source changes
- updated Doxygen config to look everywhere for doc inputs
(necessary to support new file locations)
- updated documenentation
- added CMake options for setting preprocessor definitions
in order to support discouraging modification of `lbalConfig.hpp`
