+++ docs/CMakeLists.txt
@@ -30,7 +30,7 @@ file (
 )
 
 set (DOXYGEN_INPUT_PATHS "${PROJECT_SOURCE_DIR}/include")
-set (DOXYGEN_EXCLUDE_PATHS "${PROJECT_SOURCE_DIR}/include/lucenaBAL/details")
+set (DOXYGEN_EXCLUDE_PATHS "")
 set (DOXYGEN_PATH_STRIP_PREFIX "${PROJECT_SOURCE_DIR}")
 set (DOXYGEN_OUTPUT_PATH "${CMAKE_CURRENT_BINARY_DIR}/docs/doxygen")
 set (DOXYGEN_EXAMPLES_PATH "${PROJECT_SOURCE_DIR}/examples")
