+++ include/lucenaBAL/details/compilers/lbalCompilerAppleClang.hpp
@@ -25,7 +25,7 @@
 
 
 //	lbal
-#include <lucenaBAL/lbalConfig.hpp>
+#include <lucenaBAL/details/lbalConfig.hpp>
 
 
 /*------------------------------------------------------------------------------
