+++ include/lucenaBAL/details/lbalCompilerSetup.hpp
@@ -33,7 +33,7 @@
 
 
 //	lbal
-#include <lucenaBAL/lbalConfig.hpp>
+#include <lucenaBAL/details/lbalConfig.hpp>
 
 
 /*------------------------------------------------------------------------------
