+++ include/lucenaBAL/details/lbalDefinitionTests.hpp
@@ -17,7 +17,7 @@
 
 
 //	lbal
-#include <lucenaBAL/lbalConfig.hpp>
+#include <lucenaBAL/details/lbalConfig.hpp>
 
 
 /*------------------------------------------------------------------------------
