+++ include/lucenaBAL/details/platforms/lbalPlatformPOSIX.hpp
@@ -25,7 +25,7 @@
 
 
 //	lbal
-#include <lucenaBAL/lbalConfig.hpp>
+#include <lucenaBAL/details/lbalConfig.hpp>
 
 
 /*------------------------------------------------------------------------------
