+++ include/lucenaBAL/lucenaBAL.hpp
@@ -110,5 +110,5 @@
 	@defgroup lbal_decorators Linker Decorators
 */
 
-#include <lucenaBAL/lbalConfig.hpp>
+#include <lucenaBAL/details/lbalConfig.hpp>
 #include <lucenaBAL/details/lbalFeatureSetup.hpp>
